+++ packages/rekapager-doctrine-collections-adapter/src/Internal/SelectableKeysetItem.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Rekalogika\Rekapager\Doctrine\Collections\Internal;
 
 use Doctrine\Common\Collections\Expr\ClosureExpressionVisitor;
+use Rekalogika\Contracts\Rekapager\Exception\NullBoundaryValueException;
 use Rekalogika\Rekapager\Keyset\Contracts\KeysetItemInterface;
 
 /**
@@ -65,6 +66,10 @@ final class SelectableKeysetItem implements KeysetItemInterface
         foreach ($this->boundaryProperties as $property) {
             /** @var mixed $value */
             $value = ClosureExpressionVisitor::getObjectFieldValue($this->objectOrArrayValue, $property);
+
+            if ($value === null) {
+                throw new NullBoundaryValueException(sprintf('The property "%s" of the value "%s" is a boundary value of this pagination, but it is found to be null. Null value in a boundary value is not supported.', $property, get_debug_type($value)));
+            }
 
             /** @psalm-suppress MixedAssignment */
             $result[$property] = $value;